+++ src/routes/game-jam/+layout.ts
@@ -1,0 +1,7 @@
+import type { Post } from '$lib/types'
+
+export async function load({ fetch }) {
+	const response = await fetch('../api/content')
+	const posts: Post[] = await response.json()
+	return { posts }
+}
